refactor: optimize sidebar component logic
