+++ apps/frontend/components/tatami_sidebar/app-sidebar.tsx
@@ -1,7 +1,5 @@
 "use client";
 
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
 import {
   Sidebar,
   SidebarContent,
@@ -12,218 +10,12 @@ import {
   SidebarMenuButton,
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
-import type { Model } from "@/types/models";
-import {
-  ChevronDown,
-  ChevronRight,
-  Database,
-  LayoutTemplate,
-  Pencil,
-  Trash2,
-} from "lucide-react";
-import { useEffect, useState } from "react";
-import { PropertyItem } from "../models/PropertyItem";
 import { ModelSidebarSection } from "./model-sidebar-section";
-
-const staticMenuItems = [
-  { id: "models", label: "Models", icon: Database },
-  { id: "metrics", label: "Metrics", icon: LayoutTemplate },
-];
-
-const initialDynamicContent = {
-  layers: [
-    { id: "layer1", label: "Layer 1" },
-    { id: "layer2", label: "Layer 2" },
-    { id: "layer3", label: "Layer 3" },
-  ],
-  settings: [
-    { id: "account", label: "Account Settings" },
-    { id: "preferences", label: "Preferences" },
-    { id: "theme", label: "Theme" },
-  ],
-};
+import { useSidebar } from "./useSidebar";
 
 export function AppSidebar() {
-  const [selectedOption, setSelectedOption] = useState<string | null>(null);
-  const [dynamicContent, setDynamicContent] = useState(initialDynamicContent);
-  const [models, setModels] = useState<Model[]>([]);
-  const [editingModelId, setEditingModelId] = useState<string | null>(null);
-  const [editModelName, setEditModelName] = useState<string>("");
-
-  useEffect(() => {
-    fetch("/api/models")
-      .then((res) => res.json())
-      .then((data) => setModels(data.models))
-      .catch((err) => console.error("Error loading models:", err));
-  }, []);
-
-  const saveModelsToJson = async (updatedModels: Model[]) => {
-    await fetch("/api/models", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ models: updatedModels }),
-    });
-  };
-
-  const toggleOption = (optionId: string) => {
-    if (selectedOption === optionId) {
-      setSelectedOption(null);
-    } else {
-      setSelectedOption(optionId);
-    }
-  };
-
-  const addModel = () => {
-    const newModel: Model = {
-      id: `model_${Date.now()}`,
-      name: `New Model ${models.length + 1}`,
-      expanded: true,
-      properties: [
-        { id: `prop_${Date.now()}`, name: "", dataType: "u32", isKey: true },
-      ],
-    };
-
-    setModels((prevModels) => {
-      const updatedModels = [...prevModels, newModel];
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const addProperty = (modelId: string) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId
-          ? {
-              ...model,
-              properties: [
-                ...model.properties,
-                {
-                  id: `prop_${Date.now()}`,
-                  name: "",
-                  dataType: "u32",
-                  isKey: model.properties.length === 0,
-                },
-              ],
-            }
-          : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const deleteProperty = (modelId: string, propertyId: string) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId
-          ? {
-              ...model,
-              properties: model.properties.filter((p) => p.id !== propertyId),
-            }
-          : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const updatePropertyName = (
-    modelId: string,
-    propertyId: string,
-    name: string,
-  ) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId
-          ? {
-              ...model,
-              properties: model.properties.map((p) =>
-                p.id === propertyId ? { ...p, name } : p,
-              ),
-            }
-          : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const updatePropertyDataType = (
-    modelId: string,
-    propertyId: string,
-    dataType: string,
-  ) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId
-          ? {
-              ...model,
-              properties: model.properties.map((p) =>
-                p.id === propertyId ? { ...p, dataType } : p,
-              ),
-            }
-          : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const updatePropertyKey = (
-    modelId: string,
-    propertyId: string,
-    isKey: boolean,
-  ) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId
-          ? {
-              ...model,
-              properties: model.properties.map((p) =>
-                p.id === propertyId ? { ...p, isKey } : p,
-              ),
-            }
-          : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const toggleModelExpansion = (modelId: string) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId ? { ...model, expanded: !model.expanded } : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
-
-  const updateModelName = (modelId: string, name: string) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.map((model) =>
-        model.id === modelId ? { ...model, name } : model,
-      );
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-    setEditingModelId(null);
-  };
-
-  const startEditingModelName = (modelId: string, currentName: string) => {
-    setEditingModelId(modelId);
-    setEditModelName(currentName);
-  };
-
-  const deleteModel = (modelId: string) => {
-    setModels((prevModels) => {
-      const updatedModels = prevModels.filter((model) => model.id !== modelId);
-      saveModelsToJson(updatedModels);
-      return updatedModels;
-    });
-  };
+  const { selectedOption, dynamicContent, staticMenuItems, toggleOption } =
+    useSidebar();
 
   return (
     <div className="flex h-[calc(100vh-64px)]">
